use-cases/ai/orchestrator/pyproject.toml
[project]
name = "a2a-orchestrator"
version = "0.1.0"
description = "A2A multi-agent orchestrator"
requires-python = ">=3.12"
dependencies = [
    "a2a-agents",
    "a2a-sdk[http-server]>=0.2.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=0.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "structlog>=24.0.0",
]

[tool.uv.sources]
a2a-agents = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]
